codebase/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gated_nwp"
version = "0.1.0"
description = "Metadata-aware gated attention for L2 next-word prediction"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Anonymous", email = "anon@example.com" }]
keywords = ["nlp", "second-language-acquisition", "gated-attention", "cognitive-modelling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.1",
    "transformers>=4.40",
    "datasets>=2.18",
    "accelerate>=0.28",
    "numpy>=1.26",
    "pandas>=2.1",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "click>=8.1",
    "spacy>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.9",
    "pre-commit>=3.7",
    "types-PyYAML",
    "types-tqdm",
]
viz = [
    "matplotlib>=3.8",
    "seaborn>=0.13",
]
tracking = [
    "wandb>=0.17",
]

[project.urls]
Homepage = "https://github.com/berstearns/public-metadata-aware-nwp-in-sla"
Paper = "https://github.com/berstearns/public-metadata-aware-nwp-in-sla/blob/main/main.pdf"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [".venv", "runs", "data-cache"]

[tool.ruff.lint]
select = [
    "E", "F", "W",        # pyflakes + pycodestyle
    "I",                  # isort
    "N",                  # pep8-naming
    "UP",                 # pyupgrade
    "B",                  # bugbear
    "SIM",                # simplify
    "TID",                # tidy-imports
    "RUF",                # ruff-specific
]
ignore = [
    "E501",   # line length — handled by formatter
    "RUF002", # unicode in docstrings (we write σ, ×, ⊙ on purpose)
    "N812",   # torch.nn.functional as F is canonical
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["N802", "N803"]   # allow non-pep8 names in tests
"scripts/*" = ["E402"]         # late imports after sys.path tweaks

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
ignore_missing_imports = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "smoke: end-to-end smoke tests requiring tokenisers / small models",
    "slow: tests that take more than a few seconds",
]
addopts = "-ra --strict-markers"
